Move DryMassProperties to common.mass_properties
DryMassProperties is a layer-neutral value object (no model dependencies),
not a thrust-chamber concept. Relocate it to machwave/common/mass_properties
so it sits below core and models, letting future component-level dry-mass
aggregation (tanks, feed systems) and any core numeric routine reference it
without importing models.

It is not re-exported from the thrust_chamber package; consumers import it
from machwave.common.mass_properties directly.
