+++ machwave/models/thrust_chamber/__init__.py
@@ -1,6 +1,5 @@
 from machwave.models.thrust_chamber.base import (
     BiliquidEngineThrustChamber,
-    DryMassProperties,
     SolidMotorThrustChamber,
     ThrustChamber,
 )
@@ -15,7 +14,6 @@ from machwave.models.thrust_chamber.nozzle import Nozzle
 
 __all__ = [
     "BiliquidEngineThrustChamber",
-    "DryMassProperties",
     "SolidMotorThrustChamber",
     "ThrustChamber",
     "BipropellantInjector",
